+++ applications/amathea/src/features/dataviews/dataviews.selectors.ts
@@ -9,13 +9,9 @@ import { selectAllDataviews, selectDrafDataviewSource } from './dataviews.slice'
 export const selectCurrentWorkspaceDataviews = createSelector(
   [selectAllDataviews, selectCurrentWorkspace, selectAllDatasets],
   (dataviews, workspace, datasets) => {
-    return dataviews
-      .filter((dataview) => workspace?.dataviews?.map((d) => d.id).includes(dataview.id))
-      .map((dataview) => {
-        const datasetId = dataview.datasets?.length ? dataview.datasets[0].id : ''
-        const dataset = datasets.find((dataset) => dataset.id === datasetId)
-        return { ...dataview, dataset }
-      })
+    return dataviews.filter((dataview) =>
+      workspace?.dataviews?.map((d) => d.id).includes(dataview.id)
+    )
   }
 )
 
@@ -26,24 +22,49 @@ export const selectCurrentWorkspaceDataviewsResolved = createSelector(
   }
 )
 
-export const selectDatasetOptionsBySource = createSelector(
-  [selectAllDatasets, selectDrafDataviewSource, getUserId],
-  (datasetOptions, sourceSelected, userId) => {
-    if (!sourceSelected) return []
-    const options = datasetOptions.filter((dataset) => {
-      if (
-        dataset.source === DATASET_SOURCE_IDS.user &&
-        sourceSelected.id === DATASET_SOURCE_IDS.user
-      ) {
-        return dataset.ownerId === userId
-      }
-      return dataset.source === sourceSelected.id
+export const selectUserDatasets = createSelector(
+  [selectAllDatasets, getUserId],
+  (datasets, userId) => {
+    return datasets.filter((dataset) => {
+      return dataset.source === DATASET_SOURCE_IDS.user && dataset.ownerId === userId
     })
-    return options.map((dataset) => ({
-      id: dataset.id,
-      label: dataset.name,
-      type: dataset.type,
-      description: dataset.description,
-    }))
+  }
+)
+
+export const selectDatasetOptionsBySource = createSelector(
+  [selectUserDatasets, selectAllDataviews, selectCurrentWorkspace, selectDrafDataviewSource],
+  (userDatasets, dataviews, currentWorkspace, sourceSelected) => {
+    if (!sourceSelected) return []
+
+    if (sourceSelected.id === DATASET_SOURCE_IDS.user) {
+      return userDatasets.map((dataset) => ({
+        id: dataset.id,
+        label: dataset.name,
+        type: dataset.type,
+        description: dataset.description,
+      }))
+    }
+
+    return dataviews.flatMap((dataview) => {
+      const dataset = dataview.datasets?.find(
+        (dataset) => dataset.type === '4wings:v1' || dataset.type === 'user-context-layer:v1'
+      )
+
+      const currentWorkspaceDataviews = currentWorkspace?.dataviews.map((dataview) => dataview.id)
+      if (
+        !dataset ||
+        dataset.source !== sourceSelected.id ||
+        currentWorkspaceDataviews?.includes(dataview.id)
+      ) {
+        return []
+      }
+      return {
+        id: dataset.id,
+        dataviewId: dataview.id,
+        label: dataset.name,
+        type: dataset.type,
+        description: dataset.description,
+      }
+    })
   }
 )