+++ applications/amathea/src/features/user/user.hook.ts
@@ -1,13 +1,11 @@
 import { useSelector, useDispatch } from 'react-redux'
 import { useCallback } from 'react'
-import GFWAPI from '@globalfishingwatch/api-client'
 import { fetchUserThunk, logoutUserThunk, isUserLogged, selectUserStatus } from './user.slice'
 
 export const useUserConnect = () => {
   const dispatch = useDispatch()
   const logged = useSelector(isUserLogged)
   const status = useSelector(selectUserStatus)
-  const token = GFWAPI.getToken()
 
   const fetchUser = useCallback(() => {
     dispatch(fetchUserThunk())
@@ -17,5 +15,5 @@ export const useUserConnect = () => {
     dispatch(logoutUserThunk())
   }, [dispatch])
 
-  return { logged, status, token, fetchUser, logoutUser }
+  return { logged, status, fetchUser, logoutUser }
 }