+++ applications/amathea/src/store.ts
@@ -10,6 +10,7 @@ import aoiReducer from './features/areas-of-interest/areas-of-interest.slice'
 import datasetsReducer from './features/datasets/datasets.slice'
 import dataviewsReducer from './features/dataviews/dataviews.slice'
 import workspacesReducer from './features/workspaces/workspaces.slice'
+import resourcesReducer from './features/resources/resources.slice'
 import userReducer from './features/user/user.slice'
 import mapReducer from './features/map/map.slice'
 import timebarReducer from './features/timebar/timebar.slice'
@@ -29,6 +30,7 @@ const rootReducer = combineReducers({
   datasets: datasetsReducer,
   dataviews: dataviewsReducer,
   workspaces: workspacesReducer,
+  resources: resourcesReducer,
   map: mapReducer,
   timebar: timebarReducer,
   location: location,