+++ applications/temporalgrid-demo/src/App.js
@@ -1,18 +1,51 @@
-import './App.css';
-import React, { useState, useMemo, useRef, useCallback } from 'react';
+import 'babel-polyfill';
+import React, {useState, useMemo, useRef, useCallback} from 'react';
+import {render} from 'react-dom';
 import MapGL from '@globalfishingwatch/react-map-gl';
 import { DateTime } from 'luxon'
 import { Generators } from '@globalfishingwatch/layer-composer';
 import { useLayerComposer, useDebounce } from '@globalfishingwatch/react-hooks';
 import TimebarComponent from '@globalfishingwatch/timebar';
+import Tilesets from './Tilesets';
 
-// const DEFAULT_TILESET = 'carriers_v8'
-const DEFAULT_TILESET = 'fishing_v3'
+import './App.css'
+import '@globalfishingwatch/mapbox-gl/dist/mapbox-gl.css'
 
-function App() {
+export const DEFAULT_TILESETS = [
+  {
+    // tileset: 'carriers_v8',
+    tileset: 'fishing_v4',
+    // filter: ''
+    filter: "flag='ESP'",
+    active: true
+  },
+  {
+    tileset: 'fishing_v4',
+    filter: "flag='FRA'",
+    active: false
+  },
+  {
+    tileset: 'fishing_v4',
+    filter: "flag='ITA'",
+    active: false
+  },
+  {
+    tileset: 'fishing_v4',
+    filter: "flag='GBR'",
+    active: false
+  },
+  {
+    tileset: 'fishing_v4',
+    filter: "flag='PRT'",
+    active: false
+  }
+]
+
+
+export default function App() {
   const [viewport, setViewport] = useState({
-    longitude: -17.3163661,
-    latitude: 16.3762596,
+    longitude: -6,
+    latitude: 47,
     zoom: 4.6424032
   });
 
@@ -22,8 +55,8 @@ function App() {
   })
   const debouncedTime = useDebounce(time, 1000)
 
-  const [tileset, setTileset] = useState(DEFAULT_TILESET)
-  const [currentTileset, setCurrentTileset] = useState(DEFAULT_TILESET)
+  const [tilesets, setTilesets] = useState(DEFAULT_TILESETS)
+  const [combinationMode, setCombinationMode] = useState('add')
 
   const [showBasemap, setShowBasemap] = useState(true)
   const [animated, setAnimated] = useState(true)
@@ -34,6 +67,8 @@ function App() {
   const [showInfo, setShowInfo] = useState(false)
 
   const [isPlaying, setIsPlaying] = useState(false)
+
+  const [isLoading, setLoading] = useState(false)
 
   const layers = useMemo(
     () => {
@@ -46,24 +81,33 @@ function App() {
       }
 
       if (animated) {
+
         let geomType = geomTypeMode
         if (geomType === 'blobOnPlay') {
           geomType = (isPlaying) ? 'blob' : 'gridded'
         }
+        const activeTilesets = tilesets.filter(t => t.active)
+        const colorRamps = (activeTilesets.length === 1 || combinationMode === 'add')
+          ? ['presence']
+          : ['sky', 'magenta', 'yellow', 'salmon', 'green'].slice(0, activeTilesets.length)
         generators.push({
           id: 'heatmap-animated',
           type: Generators.Type.HeatmapAnimated,
-          tileset,
+          tilesets: activeTilesets.map(t => t.tileset),
+          filters: activeTilesets.map(t => t.filter),
           debug,
           debugLabels,
           geomType,
-          tilesAPI: 'https://fourwings.api.dev.globalfishingwatch.org/v1'
+          // tilesAPI: 'https://fourwings.api.dev.globalfishingwatch.org/v1'
+          tilesAPI: ' https://fourwings-tile-server-jzzp2ui3wq-uc.a.run.app/v1/datasets',
+          combinationMode,
+          colorRamps,
         })
       } else {
         generators.push({
           id: 'heatmap',
           type: Generators.Type.Heatmap,
-          tileset,
+          tileset: tilesets.tileset[0],
           visible: true,
           geomType: 'gridded',
           serverSideFilter: undefined,
@@ -76,7 +120,7 @@ function App() {
 
     return generators
   },
-    [viewport, showBasemap, debug, debugLabels, tileset, geomTypeMode, isPlaying]
+    [animated, viewport, showBasemap, debug, debugLabels, tilesets, geomTypeMode, isPlaying, combinationMode]
   );
 
   // TODO switch between debounced/immediate/throttled time when using animated
@@ -85,6 +129,12 @@ function App() {
   const mapRef = useRef(null)
   if (mapRef && mapRef.current) {
     mapRef.current.getMap().showTileBoundaries = debug
+    mapRef.current.getMap().on('idle', () =>  {
+      setLoading(false)
+    })
+    mapRef.current.getMap().on('dataloading', () =>  {
+      setLoading(true)
+    })
   }
 
   const onMapClick = useCallback((e) => {
@@ -95,6 +145,7 @@ function App() {
 
   return (
     <div className="container">
+      {isLoading && <div className="loading">loading</div>}
       <div className="map">
         <MapGL
           {...viewport}
@@ -104,6 +155,7 @@ function App() {
           mapStyle={style}
           onViewportChange={nextViewport => setViewport(nextViewport)}
           onClick={onMapClick}
+          onData={() => { console.log('loafed') }}
         />
       </div>
       <div className="timebar">
@@ -123,10 +175,8 @@ function App() {
         </TimebarComponent>
       </div>
       <div className="control-buttons">
-        <fieldset>
-          <input type="text" value={currentTileset} onChange={(event) => setCurrentTileset(event.target.value)} />
-          <button onClick={() => setTileset(currentTileset)}>ok</button>
-        </fieldset>
+        <Tilesets onChange={(tilesets, combinationMode) => { setTilesets(tilesets); setCombinationMode(combinationMode) }} />
+        <hr />
         <fieldset>
           <input type="checkbox" id="showBasemap" checked={showBasemap} onChange={(e) => {
             setShowBasemap(e.target.checked)
@@ -174,4 +224,6 @@ function App() {
   );
 }
 
-export default App;
+export function renderToDom(container) {
+  render(<App />, container);
+}