+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @globalfishingwatch/layer-composer
 
+## 2.8.1
+
+### Patch Changes
+
+- e6f5aa7: Update enpdoints ids
+
+## 2.8.0
+
+### Minor Changes
+
+- a566ff6: Layer component
+
+## 2.7.1
+
+### Patch Changes
+
+- 8912bb0: fix temporal-grid demo
+
+## 2.7.0
+
+### Minor Changes
+
+- ba5064b: fix temporalgrid filters
+
 ## 2.6.1
 
 ### Patch Changes